move files in package
