+++ RSSReader/app/src/main/java/ru/ifmo/md/lesson5/rssreader/utils/ChannelsLoader.java
@@ -1,8 +1,10 @@
-package ru.ifmo.md.lesson5.rssreader;
+package ru.ifmo.md.lesson5.rssreader.utils;
 
 import android.content.Context;
 import android.content.CursorLoader;
 import android.database.Cursor;
+
+import ru.ifmo.md.lesson5.rssreader.RSSManager;
 
 /**
  * Created by Nikita Yaschenko on 19.10.14.
